Provide CQL interval operators in queries. Use generators to yield query results.
This commit rewrites the mechanism how queries are executed. First of all, each
query string is preprocessed during the construction of a CorpusQuery instance:
if the query string specifies an interval operator, a list of new query strings
are generated that expand the operator so that all requested numbers of repetitions
are explicitly stated in a separate query string. A subquery is constructed for
each of these strings, and these subqueries are stored in a list for the main
query. Session.run_queries() now checks if a query has subqueries. If so, they
are sequentally executed instead of the main query.

Due to this change, Corpus.run_queries() has been rewritten. It is now called
Corpus.yield_query_results(). It is now a generator that yields entries from
the SQL query cursor.