Add new corpus building class. Add installers for some corpora.
In order to make the corpus building process more generic, the new class
BaseCorpusBuilder is introduced in the file corpusbuilder.py. Corpus
installers can inherit this class, and overwrite all corpus-specific
parts. One new feature of this class is that it can write a Python
library to a specified path.

Writing an installer for a new corpus should thus eintail only the
following steps:

- write an descendent of the BaseCorpusBuilder class
- include all corpus-specific labels in the class
- add all table descriptions
- add corpus-specific code
- deploy the Python code as an installer.

The first three building tools to use this are install_generic.py,
install_buckeye.py, and install_bnc.py. They are also heavily commented so
that they can serve as a guideline for people who want to write installers
for their own corpora.