Make the option to output source information more flexible. Fix source filters.
Users can now use the --source/-s argument to choose which source information
columns will be included in the output. --source ALL will include all columns
that are specified in the Resource. Every attribute of Resource with a name
that starts with '.source_info_' is considered a source info column name.

This argument can be specified multiple times to select e.g. the date and the
year, like so:

coquery --source Date --source Year

This mechanism should be added to all output controlling arguments, e.g.
for --speaker.