CoqTextTag, CoqFilterTag and CoqTagBox classes.
The CoqTagBox class is a Qt widget that can be used to manage tags. The widget
keeps track of the entered text strings, which occur as "tags" in a flowing
layout above the text edit. The tags can be rearranged, and also deleted.

In Coquery, these classes are used to provide an intuitive way to enter query
filters.