Make the experimental mode faster and more versatile. Add context renderer.
The nexperimental mode does no longer attempt to retrieve the context also in
a single query, this was awfully slow, and as MySQL/MariaDB can only have up
to 61 JOINs in a single query, context span was effectively limited to
14 words, which is rather poor. Now, the contexts are provided for each
query result in yield_query_results() separately.

The initial context renderer is still very basic, but it should be possible
to write fairly complex renderers that can produce complex renditions of
the original corpus sources if the corpus provides this information. For
example, the BNC has information about headlines, lists, and so on. This
information is currently ignored, but it could be incorporated into the
database as separate annotation tables.

This may turn out to be an extremely intersting feature.