Add new query mode: QUERY_MODE_COLLOCATIONS
This is a rather big commit: it adds a completely new query mode that allows
to retrieve collocates and their collocate frequencies for a query. It also
calculates the mutual information score.

The algorithm first expands the query with '*' wildcards to the left and the
right, and runs a normal token query. Then, all words in the collocation
neighbourhood are counted, and their frequency is obtained. The query then
produces one line of output for each collocate.

There is still some room for improvement. The frequencies may be obtained
for all collocates in one query, potentially speeding up their retrieval.
And the collocate measures are incorrect if result filters are used (see
Issue #21).