Provide a framework for data visualization.
The class Visualizer() provides the framework for new visualizations. Its
method draw() can be overridden to create a new plot type, and the method
setup_figure() can be used to define the grid layout required for this plot
type. Visualizer classes have the class attribute visualize_frequency. If
True, the class provides a plot type that can be used to display the result
of frequency queries. If False, it provides a plot type that can be used to
display token queries (either DISTINCT or ALL).

Visualizer() also contains some interfaces to the result table, but these
functions may be moved to an independent class that provides some dataframe-
like interfaces.

The second base class is VisualizerDialog(), which can be used to provide a
dialog window for visualizations. This dialog makes use of the Matplotlib
navigation toolbar for zooming, panning, and saving. It also provides the
capability to freeze and unfreeze the current visualization.

A visualization dialog with a specific visualization type can be opened by
calling the static method Plot().