Add performance improvements. Add subquery order heuristic.
This commit provides a few changes that should increase query performance.
First of all, it reduces the number of fields that are included in a query,
which most notably reduces the number of joins that are necessary for many
queries. Most importantly, queries don't include the fields required to
represent the token specifications anymore (this was a left-over from when
there was no call to get_matching_wordids() for each token at the beginning
of queries). Second, coq_invisible_origin_id was not pulled in from the
corpus table, but from a joined table.

The second change concerns the order in which self-joins of the corpus table
are done for multi-word queries. The default was to self-join the corpus
table sequentially from first to last token, but this caused horrible
performances if the query string started with tokens that matched a large
number of rows, most strikingly '*' tokens.

The new method now self-joins the corpus table on the basis of a heuristic
that takes into account the number of characters and the presence of
asterisks '*' in the query token, which hopefully leads to much smaller
self-joins. This heuristic may still need improvement, though.