Make use of query item mappings.
This commit very much changes the query string construction logic. It does
not attempt anymore to guess the lemma, transcript, and part-of-speech
resources, but uses the query item mappings for that. This change has
two advantages.

First, the code becomes much more concise and readable; some redundancies
 are removed.

Second, and perhaps more importantly, the database layout is now fully
flexible. Before the change, it was assumed that e.g. the part-of-speech
information is stored either in the corpus or in the word table, or in a
table named 'pos' that is linked to the word table. Now, the POS tag can
be in any available resource feature of the corpus. If it is not in one
of the default resource features (see set_query_items() from
corpusbuilder.py for the defaults), the corpus installer has just to
include a call to map_query_item() to specify where the POS tag resides
(see coq_install_celex.py for an example).