Rewrite mechanism for external links.
This is the first step in a complete overhaul of the way how external tables
are linked. There is now a new Link class that implements table joins in a
more consistent way than before. This will eventually provide a fix for
issues #176 and #92.

A visible change is that links are now persistent. They are not only restored
when you change back to a corpus to which an external was linked, they are
also restored on start-up.