Add dyplr-inspired data management
This commit introduces a new way of managing the data columns. Now, query
results are handled in a way that is similar to how you would manage them
in R with dyplr.

It introduces the concept of data engines, which replace the functionality
of the different queries, and dedicated Function objects, which operate
directly on the data frames. They allow calculations, reg exs, and all
other kinds of nice features.